+++ apps/docs/styles/main.scss
@@ -166,9 +166,9 @@ pre[class*='language-'] {
 }
 
 // ToC styles
-.toc__menu-item--active {
-  color: hsl(var(--brand-default)) !important;
-}
+// .toc__menu-item--active {
+//   color: hsl(var(--brand-default)) !important;
+// }
 
 .video-container {
   position: relative;
