+++ packages/ui-patterns/Toc/index.ts
@@ -3,14 +3,12 @@ import { ScrollProvider, TOCItem, type AnchorProviderProps, AnchorProvider } fro
 import { Toc, TOCScrollArea, TOCItems } from './toc'
 import { TocThumb } from './toc-thumb'
 import { type TOCThumb } from './toc-thumb'
-import TocInsetItems from './toc-inset'
 
 export {
   AnchorProvider,
   Toc,
   TocPrimitive,
   TocThumb,
-  TocInsetItems,
   TOCItem,
   TOCScrollArea,
   TOCItems,
